refactor(web): make shell/ filenames match their exports
Every file in components/shell/ now has a basename matching its primary
export. The sidebar chrome forms a uniform sidebar-*/Sidebar family and the
page container a page-*/Page one, so a reader can grep either from a symbol.

- adaptive-sidebar.tsx/AdaptiveShellSidebar -> sidebar-adaptive.tsx/SidebarAdaptive
- floating-trigger.tsx -> sidebar-floating-trigger.tsx
- dropdown-menu.tsx -> sidebar-dropdown-menu.tsx (no longer shadows ui/dropdown-menu)
- user-menu.tsx -> sidebar-user-menu.tsx
- content.tsx (PageShell + PageHeader) split into page-shell.tsx + page-header.tsx

Updates every importer plus the design skill, manage/dashboard.mdx, and the
segregation audit doc.
