+++ web/next/src/components/shell/sidebar-adaptive.tsx
@@ -6,7 +6,7 @@ import { Sidebar } from "@/components/ui/sidebar"
 import { isDocsPath } from "@/lib/docs"
 
 // The shell sidebar is an icon rail in the app, but offcanvas inside the docs area: icon-less doc trees don't collapse to a rail cleanly, whereas offcanvas hides fully and pairs with the floating edge tab.
-export function AdaptiveShellSidebar({ children }: { children: React.ReactNode }) {
+export function SidebarAdaptive({ children }: { children: React.ReactNode }) {
   const collapsible = isDocsPath(usePathname()) ? "offcanvas" : "icon"
 
   return <Sidebar collapsible={collapsible}>{children}</Sidebar>
